Moved loading measurements to session details controller
